Added invalid input tests, corrected bug in offset validator
Correctly handles null inputs on start address now.

+++ src/utilities/eui_offset_validation.h
@@ -6,7 +6,7 @@
 void
 validate_offset_range(  uint16_t base, 
                         uint16_t offset, 
-                        uint16_t type_bytes,
+                        uint8_t  type_bytes,
                         uint16_t size,
                         uint16_t *start,
                         uint16_t *end );
